fix: use u32 for Discord embed color
We were using u8 for Discord embed colors, which isn't enough since
colors require 3-4 bytes. u32 is much better suited for this, let's use
it.

+++ src/main.rs
@@ -84,7 +84,7 @@ fn format_discord_message(activity: &Activity) -> WebhookMessage {
             .and_then(|i| {
                 i.color
                     .as_ref()
-                    .map(|c| u8::from_str_radix(c.trim_start_matches('#'), 16).ok())
+                    .map(|c| u32::from_str_radix(c.trim_start_matches('#'), 16).ok())
             })
             .flatten(),
